% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{generar_radares}
\alias{generar_radares}
\title{Generar datos para radares}
\usage{
generar_radares(anio, sector, filtro = NULL)
}
\arguments{
\item{filtro}{}
}
\description{
Generar datos para radares
}
